+++ src/components/index.ts
@@ -105,6 +105,6 @@ export { showToast } from './Status/Toast/utils'
 
 export { MapControls } from './presets/MapControls'
 export { IconMarkers } from './presets/IconMarkers'
-export { SSOButtons } from './presets/SSOButtons'
+export { SSOButtons } from './Forms/Actions/Button/preset/SSOButtons'
 
 // -- Presets -- //